requirements.txt
# Telegram API
telethon==1.36.0

# Database
aiosqlite==0.19.0

# HTTP requests
aiohttp==3.9.1
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date and time
python-dateutil==2.8.2
pytz==2023.3

# Logging
coloredlogs==15.0.1

# CLI
click==8.1.7
